lsコマンドの実装&クラス化
